requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
beautifulsoup4
requests
